+++ misc/read.c
@@ -96,7 +96,7 @@ ssize_t readall_interrupt(int fd, void *buf, size_t count, int interruptfd)
     ret = read(fd, cbuf + bytes_read, count - bytes_read);
     if (ret > 0)
     {
-      bytes_read += ret;
+      bytes_read += (size_t)ret;
     }
     if (ret == 0)
     {
@@ -111,7 +111,7 @@ ssize_t readall_interrupt(int fd, void *buf, size_t count, int interruptfd)
       return ret;
     }
   }
-  return bytes_read;
+  return (ssize_t)bytes_read;
 }
 
 ssize_t readall(int fd, void *buf, size_t count)
@@ -124,7 +124,7 @@ ssize_t readall(int fd, void *buf, size_t count)
     ret = read(fd, cbuf + bytes_read, count - bytes_read);
     if (ret > 0)
     {
-      bytes_read += ret;
+      bytes_read += (size_t)ret;
     }
     if (ret == 0)
     {
@@ -139,5 +139,5 @@ ssize_t readall(int fd, void *buf, size_t count)
       return ret;
     }
   }
-  return bytes_read;
+  return (ssize_t)bytes_read;
 }
